* origin/main:
  fix: update go-grpc dependency versions (#1591)
  refactor(datapath): replace Pool bitmap/vec with three-vec design (#1496)

+++ data-plane/core/service/src/service.rs
@@ -489,7 +489,7 @@ impl Service {
                 // Get connection details from the connection table
                 self.message_processor
                     .connection_table()
-                    .get(conn_id as usize)
+                    .get(conn_id)
                     .map(|conn| ConnectionInfo {
                         id: conn_id,
                         remote_addr: conn.remote_addr().copied(),
@@ -763,7 +763,7 @@ mod tests {
             service
                 .message_processor
                 .connection_table()
-                .get(conn_id as usize)
+                .get(conn_id)
                 .is_none(),
             "connection should be removed after disconnect"
         );
